requirements.txt
python-dotenv==1.0.1
pandas==2.1.2
requests==2.31.0